requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
strands-agents[ollama]>=0.1.0
requests>=2.28.0
geopy>=2.4.0
